Bug fix in printing of IPv6 addresses.
Previous implementation would append an extra colon to the SB and then delete it under certain conditions. I found this logic a bit convoluted, so I changed the implementation to append only what is needed and exit fast. I also renamed some of the variables to make things more clear.